requirements.txt
aiohttp~=3.7.3
PyJWT~=2.0.1
httpx~=0.16.1
uvicorn~=0.13.3
fastapi~=0.63.0